modified   PNXLcode/cgitraces.cgi modified   PNXLcode/defaults.ini modified   PNXLcode/findsettings modified   PNXLcode/gettraces modified   PNXLcode/MCA.csv modified   PNXLcode/PixieNetConfig.cpp modified   PNXLcode/PixieNetConfig.h modified   PNXLcode/PixieNetDefs.h modified   PNXLcode/progfippi modified   PNXLcode/progfippi.c modified   PNXLcode/RS.csv modified   PNXLcode/settings.ini modified   PNXLcode/startdaq modified   PNXLcode/startdaq.c - debug eth data out

+++ PixieNetConfig.h
@@ -122,11 +122,11 @@ typedef struct PixieNetFippiConfig {
   unsigned long long DEST_MAC0;      // destination MAC for data from first K7 
   unsigned long long DEST_MAC1;      // destination MAC for data from second K7 
 
-  unsigned int DEST_IP0;             // destination IP for data from first K7 
-  unsigned int DEST_IP1;             // destination IP for data from second K7 
+  unsigned long long DEST_IP0;             // destination IP for data from first K7 
+  unsigned long long DEST_IP1;             // destination IP for data from second K7 
 
-  unsigned int SRC_IP0;              // source IP for data from first K7   (should match WR IP from PROM)
-  unsigned int SRC_IP1;              // source IP for data from second K7 
+  unsigned long long SRC_IP0;              // source IP for data from first K7   (should match WR IP from PROM)
+  unsigned long long SRC_IP1;              // source IP for data from second K7 
 
 
   /* SYS_U## : reserved parameters  */
